LPS-205919 Remove scr pojo service wiring design in AnalyticsCloudClient
