+++ modules/apps/analytics/analytics-settings-rest-impl/src/main/java/com/liferay/analytics/settings/rest/internal/resource/v1_0/CommerceChannelResourceImpl.java
@@ -17,6 +17,7 @@ import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.service.GroupService;
+import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.vulcan.dto.converter.DTOConverter;
@@ -88,6 +89,8 @@ public class CommerceChannelResourceImpl
 			_portal.getClassNameId(
 				"com.liferay.commerce.product.model.CommerceChannel")
 		};
+
+		_analyticsCloudClient = new AnalyticsCloudClient(_http);
 	}
 
 	private LinkedHashMap<String, Object> _getParams() {
@@ -96,9 +99,7 @@ public class CommerceChannelResourceImpl
 		).build();
 	}
 
-	@Reference
 	private AnalyticsCloudClient _analyticsCloudClient;
-
 	private long[] _classNameIds;
 
 	@Reference(
@@ -111,6 +112,9 @@ public class CommerceChannelResourceImpl
 
 	@Reference
 	private GroupService _groupService;
+
+	@Reference
+	private Http _http;
 
 	@Reference
 	private Portal _portal;
